feat: create a custom hook for fetching games.
components are cleaner.
