+++ src/App.tsx
@@ -16,14 +16,17 @@ function App() {
       <GridItem pl='2' area={'nav'}>
         <Navbar />
       </GridItem>
+
       <Show above='lg'>
         <GridItem pl='2' area={'aside'}>
           Aside
         </GridItem>
       </Show>
+
       <GridItem pl='2' area={'main'}>
         <GameGrid />
       </GridItem>
+      
       <GridItem pl='2' area={'footer'}>
         Footer
       </GridItem>
