+++ src/components/GameGrid.tsx
@@ -1,29 +1,10 @@
-import { useEffect, useState } from "react"
-import apiClient from "../services/api-client";
 import { Text } from "@chakra-ui/react";
+import useGames from "./hooks/useGames";
 
-interface Game {
-  id: number
-  name: string
-}
-
-interface fetchGamesResponse {
-  count: number
-  results: Game[]
-}
 
 const GameGrid = () => {
-  const [games, setGames] = useState<Game[]>([])
-  const [error, setError] = useState('');
-
-  useEffect(() => {
-    apiClient<fetchGamesResponse>('/games')
-      .then(res => setGames(res.data.results))
-      .catch(err => setError(err.message))
-  }, [])
-
-
-
+  const { games, error } = useGames()
+  
   return (
     <>
       <Text>{error && <p>{error}</p>}</Text>
